Move LoadPromotions out of OrderAdjuster namespace
Loading the currently available promotions can be quite useful outside
of the order adjuster. This moves the LoadPromotions class out of there
and into the main `SolidusPromotions` namespace so that it's clear that
this class can be used elsewhere.
